@remotion/effects: Improve blur sampling for large radii
